refactor: Memory adapter now stores additional data along with the value
We also added the ability to have a `ClockInterface` injected to the `Memory` adapter so that the item expiry is calculated based on the TTL on-write rather than on-read. This allows per-item TTL.

Since `StorageInterface#touchItem` functionality does update the TTL, the `Metadata` now has `mtime` and `ctime` where `ctime` is the original value where the item was created and `mtime` where the item was touched.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>