feature: Replace `memory_limit` with `max_items`
The `memory_limit` implementation is flawed. It does not respect actual memory consumption by cache items but the amount of memory consumed by the application itself. Therefore, an alternative was needed. `symfony/cache` is using `max_items` as well and thus, this might be a good alternative for the `Memory` adapter as well.

When there are `max_items` configured, the `Memory` adapter now removes the oldest item(s) before creating new entries (in case the max items were exceeded).

This allows projects to limit the cache entries rather than the memory consumption of cache entries.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>